Разделяет служебные функции по файлам и добавляет проверки на экземпляр класса Abstract
