+++ src/mock/trip-events-item.js
@@ -1,5 +1,5 @@
 import dayjs from "dayjs";
-import {getRandomInteger, shuffleArray} from "../util.js";
+import {getRandomInteger, shuffleArray} from "../utils/common";
 
 const EVENTS_TYPE = [`Taxi`, `Bus`, `Train`, `Ship`, `Transport`, `Drive`, `Flight`, `Check-in`, `Sightseeing`, `Restaurant`];
 const PLACES_OF_DESTINATION = [`Amsterdam`, `Geneva`, `Chamonix`, `Köln`];
